updated types.h PANumber

+++ include/defs.h
@@ -24,7 +24,7 @@ typedef uint32_t ListPosition;
 #endif
 // typedef int PAInt;
 #define BLK_SIZE ((int)4096)
-#define FIRST ((int)1)
+#define FIRST (1)
 #define LAST(N) (N)
 #define PADDING_1_BYTE(NUM) unsigned char padding[NUM] 
 #endif
